somehow fixes lag with pdf-view-mode ...
After experiencing crippling lag when resizing a window containing a pdf-view
buffer, I noticed that the very top of the pdf would start to show in the
minibuffer. What is more, a window would briefly appear containing the pdf just
before the which-key popup appeared; as soon as which-key was visible, the extra
window was gone. All of this is baffling.

After attempting to identify why this lag / window behavior doesn't happen in
doom, I find that this variable is set to -10. Imagine my surprise when setting
it fixes the which-key bug AND the lag in pdf-tools. Why?